Refactored files, added javadoc comments
