+++ app/src/main/java/de/bws/udrive/ui/main/SignupTabFragment.java
@@ -1,4 +1,4 @@
-package de.bws.udrive;
+package de.bws.udrive.ui.main;
 
 import android.content.Intent;
 import android.graphics.drawable.ColorDrawable;
@@ -18,6 +18,8 @@ import android.widget.EditText;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import de.bws.udrive.R;
+import de.bws.udrive.ui.home.HomeActivity;
 import de.bws.udrive.utilities.handler.SignUpHandler;
 import de.bws.udrive.utilities.model.SignUp;
