+++ app/src/main/java/de/bws/udrive/ui/meineFahrt/MeineFahrtFragment.java
@@ -1,12 +1,10 @@
 package de.bws.udrive.ui.meineFahrt;
 
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
@@ -16,14 +14,8 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import de.bws.udrive.R;
 import de.bws.udrive.databinding.FragmentMeinefahrtBinding;
-import de.bws.udrive.ui.start.adapter.MeineFahrtAdapter;
 import de.bws.udrive.utilities.handler.PassengerRequestsHandler;
-import de.bws.udrive.utilities.model.PassengerRequest;
 
 public class MeineFahrtFragment extends Fragment {
     private FragmentMeinefahrtBinding binding;
@@ -39,9 +31,6 @@ public class MeineFahrtFragment extends Fragment {
 
         binding = FragmentMeinefahrtBinding.inflate(inflater, container, false);
         View root = binding.getRoot();
-
-        final TextView textView = binding.textMeineFahrt;
-        meineFahrtViewModel.getText().observe(getViewLifecycleOwner(), textView::setText);
 
         refreshButton = binding.btnDriverRefresh;
         refreshButton.setOnClickListener(driverRefreshListener);
