+++ app/src/main/java/de/bws/udrive/ui/nachrichten/MessageViewHolder.java
@@ -1,4 +1,4 @@
-package de.bws.udrive.ui.start.adapter;
+package de.bws.udrive.ui.nachrichten;
 
 import android.view.View;
 import android.widget.TextView;
