--- app/src/main/java/de/bws/udrive/ui/start/fahrer.java
@@ -1,38 +1,0 @@
-package de.bws.udrive.ui.start;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-
-public class fahrer {
-
-    String title;
-    String description;
-    long createdTime;
-
-
-    // Getter and Setter
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public long getCreatedTime() {
-        return createdTime;
-    }
-
-    public void setCreatedTime(long createdTime) {
-        this.createdTime = createdTime;
-    }
-}
