+++ app/src/main/java/de/bws/udrive/utilities/APIClient.java
@@ -6,6 +6,7 @@ import retrofit2.converter.gson.GsonConverterFactory;
 /**
  * Klasse, die benötigt wird, um API Calls vorzunehmen
  *
+ * @author Lucas
  * */
 public class APIClient {
     private static Retrofit retrofit;
