+++ app/src/main/java/de/bws/udrive/utilities/model/DriveRequest.java
@@ -1,5 +1,10 @@
 package de.bws.udrive.utilities.model;
 
+/**
+ * Klasse, die Anfrage für verfügbare Fahrten repräsentiert
+ *
+ * @author Lucas, Niko
+ */
 public class DriveRequest {
     private double currentLatitude;
     private double currentLongitude;
