+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/map"
-    tools:context=".MainActivity">
+    tools:context=".ui.main.MainActivity">
 
     <ScrollView
         android:layout_width="match_parent"
