+++ app/src/main/res/layout/fragment_meinefahrt.xml
@@ -6,20 +6,6 @@
     android:layout_height="match_parent"
     tools:context=".ui.meineFahrt.MeineFahrtFragment">
 
-    <TextView
-        android:id="@+id/text_MeineFahrt"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="8dp"
-        android:layout_marginTop="8dp"
-        android:layout_marginEnd="8dp"
-        android:textAlignment="center"
-        android:textSize="20sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rvPassengerList"
