+++ app/src/main/res/layout/fragment_signup_tab.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    tools:context=".SignupTabFragment">
+    tools:context=".ui.main.SignupTabFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
